Fixes review findings
Add aditional white Cloudogu logo as svg-image. The blue and white logo
are used in certain places and may be exchanged there too.
Fix gulp-pipeline to copy and minify the svg-logos. svg and png images are
minified and copied to /dist. The svg-images do not need to be resized.
Include the second Cloudogu-svg-logo to index.

#16